distance works but still very buggy
